+++ provisioning/src/users_and_groups.js
@@ -3,24 +3,30 @@ const {
   createUser,
   createGroup,
   addUserToGroup,
-  removeUserFromGroup,
   grantAllPermissionsToUser,
+  grantGlobalPermissionToUser,
 } = require("./api");
 const log = require("./logger");
 
 const provisionUsers = async (axios, folder, organization) => {
   try {
     const users = readJsonFile(folder + "users.json");
+
     for (const user of users) {
-      log.info(`~> Adding user ${user.displayName}`);
-      await createUser(axios, user, organization);
+      log.info(`~> Adding user ${user.displayName}`); 
+      await createUser(axios, {id: user.id, displayName: user.displayName, password: user.password}, organization);
     }
 
-    await grantAllPermissionsToUser(axios, "mstein");
-    log.info("~> all Permissions granted to mstein");
-
-    await grantAllPermissionsToUser(axios, "jdoe");
-    log.info("~> all Permissions granted to jdoe");
+    const usersWithPermissions = users.filter((user) => user.permissions.length > 0);
+    for (const user of usersWithPermissions) {
+      if (user.permissions.includes("all")) {
+        log.info(`~> all Permissions granted to ${user.id}`);
+        await grantAllPermissionsToUser(axios, user.id);
+      } 
+      else {
+        await Promise.all(user.permissions.map((intent) => { log.info(`~> Granting permission ${intent} to ${user.id}`); return grantGlobalPermissionToUser(axios, intent, user.id);}));
+      }
+    }
   } catch (err) {
     if (err.code && err.code === "MAX_RETRIES") {
       log.info("Failed to provision users, max retries exceeded");
@@ -34,23 +40,12 @@ const provisionUsers = async (axios, folder, organization) => {
 
 const provisionGroups = async (axios, folder) => {
   try {
-    const users = readJsonFile(folder + "users.json");
     const groups = readJsonFile(folder + "groups.json");
-    for (const group of groups) {
-      log.info(`~> Adding group ${group.displayName}`);
-      await createGroup(axios, group);
+    await Promise.all(groups.map((group) => createGroup(axios, group)));
+    const groupsWithUsers = groups.filter((group) => group.users.length > 0);
+    for (const group of groupsWithUsers) {
+      await Promise.all(group.users.map((user) => addUserToGroup(axios, group.id, user)));
     }
-
-    log.info("~> Adding user jxavier to group 'Reviewers'");
-    await addUserToGroup(axios, "reviewers", "jxavier");
-    log.info("~> Adding user pkleffmann to group 'Reviewers'");
-    await addUserToGroup(axios, "reviewers", "pkleffmann");
-    log.info("~> Adding user dviolin to group 'Reviewers'");
-    await addUserToGroup(axios, "reviewers", "mstein");
-    log.info("~> Adding user mstein to group 'Reviewers'");
-    await addUserToGroup(axios, "reviewers", "dviolin");
-    log.info("~> Removing user dviolin from group 'Reviewers'");
-    await removeUserFromGroup(axios, "reviewers", "dviolin");
   } catch (err) {
     if (err.code && err.code === "MAX_RETRIES") {
       log.info("Failed to provision groups, max retries exceeded");